social_streamliner/requirements.txt
Flask[async]
gspread
python-telegram-bot
google-auth-oauthlib
oauth2client
python-dotenv
google-generativeai
requests
schedule
